Alternate solving process and time updates
User is now given the option to have guesses follow the wordle feedback or not. More refinement is needed, mainly marking which words are following the rules and which ones are not. The script may also need to give the user the ability to see proper answers if they aren't given as high of a score.

Also included a feature that reports how many words the script has processed and the percentage complete so the user isn't left wondering how much more time it will take to figure out the best word.