Improved options feature, created testing feature, improved status lines, cleaned code
Made the "options" feature happen all the time, also provided a list of words that were all answers as well as just good guesses.

Made test function, where the user can type in "test" and the program will decide on a random secret word and give feedback on your guesses according to the secret word.

Improved(?) the status line so that it will include estimated time to completion. Also made the status bar go away once it reached 100%.

Generally improved the flow of the code so that method 1 and 2 more easily worked side by side.